Move pathfinding to helper file
